+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+Version 0.61.0
+-------------
+
+This release contains changes from [PR #476](https://github.com/codemagic-ci-cd/cli-tools/pull/476).
+
+**Features**
+- Add `--altool-additional-arguments` option to the `app-store-connect publish` action. This allows passing extra CLI arguments directly to `altool` during the app upload process.
+
+**Docs**
+- Update docs for `app-store-connect publish`.
+
 Version 0.60.2
 -------------